requirements.txt
streamlit
pandas
joblib
python-dotenv
google-generativeai
requests
Pillow
functionality
